Move path resolver to its own package
