+++ pkg/security/resolvers/path/errors.go
@@ -6,7 +6,7 @@
 //go:build linux
 // +build linux
 
-package resolvers
+package path
 
 import "fmt"
